Week 4/Writing Practice from a Picture/SentencefromPicture/requirements.txt
# Web Framework
Flask>=2.3.3

# Image Processing
Pillow>=10.0.0

# OpenAI Integration
openai==0.28.1

# Environment Variables
python-dotenv>=1.0.0
